@@ -16,69 +16,70 @@
   "scripts": {
     "test": "jest",
     "start": "node server.js",
-    "build": "webpack --config conf/webpack-prod.js  --progress --profile --colors; echo 'copying to electron/build..'; cp -r dist/* ../electron/build/; echo 'done.'"
+    "build": "NODE_ENV=production webpack --config conf/webpack-prod.js  --progress --profile --colors; echo 'copying to electron/build..'; cp -r dist/* ../electron/build/; echo 'done.'"
   },
   "dependencies": {
-    "alt": "~0.17.1",
-    "autoprefixer-loader": "^2.0.0",
-    "babel": "~5.8.21",
-    "babel-core": "~5.8.22",
-    "babel-loader": "~5.3.2",
+    "alt": "^0.17.8",
     "bignumber.js": "^2.1.0",
-    "classnames": "^2.1.1",
+    "classnames": "^2.2.0",
     "cookies-js": "^1.2.1",
-    "counterpart": "^0.16.7",
-    "css-loader": "^0.12.0",
+    "counterpart": "^0.16.10",
     "event-emitter": "^0.3.4",
-    "extract-text-webpack-plugin": "^0.8.0",
-    "file-loader": "^0.8.1",
+    "git-rev-sync": "^1.4.0",
     "html-loader": "^0.3.0",
-    "immutable": "~3.7.2",
+    "immutable": "^3.7.5",
     "indexeddbshim": "^2.0.4",
     "intl": "^1.0.0",
     "jdenticon": "git+https://github.com/cryptonomex/jdenticon.git",
     "js-sha256": "^0.2.3",
-    "json-loader": "^0.5.1",
-    "keyboardjs": "^0.4.2",
     "lodash": "^3.10.1",
     "lzma": "2.1.6",
     "node-fetch": "^1.3.1",
-    "node-sass": "^3.0.0",
-    "perfect-scrollbar": "^0.6.2",
+    "perfect-scrollbar": "^0.6.7",
     "react": "0.13.3",
     "react-foundation-apps": "git+http://github.com/valzav/react-foundation-apps.git",
     "react-highcharts": "^2.1.0",
-    "react-hot-loader": "~1.2.7",
     "react-interpolate-component": "^0.7.1",
     "react-intl": "^1.2.0",
     "react-json-inspector": "5.0.3",
-    "react-notification-system": "^0.1.11",
+    "react-notification-system": "^0.1.17",
     "react-router": "0.13.3",
     "react-tools": "~0.13.2",
-    "react-tooltip": "^0.6.2",
+    "react-tooltip": "^0.6.4",
     "react-translate-component": "^0.9.0",
     "remarkable-loader": "^0.1.0",
-    "sass-loader": "^1.0.1",
-    "style-loader": "^0.12.2",
-    "svg-inline-loader": "^0.1.0",
     "svgo": "^0.5.1",
-    "svgo-loader": "^1.1.0",
-    "tcomb": "1.1.0",
+    "tcomb": "2.5.2",
     "url-loader": "^0.5.6",
-    "webpack": "^1.9.6",
-    "webpack-dev-server": "^1.8.2",
     "whatwg-fetch": "^0.9.0"
   },
   "devDependencies": {
-    "babel-eslint": "^3.1.1",
-    "babel-jest": "^5.2.0",
+    "autoprefixer-loader": "^3.1.0",
+    "babel-core": "^5.8.33",
+    "babel-eslint": "^4.1.4",
+    "babel-jest": "^5.3.0",
+    "babel-loader": "^5.3.3",
+    "babel-plugin-react-transform": "^1.1.1",
     "clean-webpack-plugin": "^0.1.3",
     "coffee-loader": "^0.7.2",
-    "eslint": "^0.20.0",
-    "eslint-plugin-react": "^2.2.0",
+    "css-loader": "^0.22.0",
+    "eslint": "^1.9.0",
+    "eslint-plugin-react": "^3.8.0",
+    "extract-text-webpack-plugin": "^0.9.1",
     "file-loader": "^0.8.4",
     "jest-cli": "~0.4.1",
+    "json-loader": "^0.5.3",
+    "node-sass": "^3.4.1",
+    "react-transform-catch-errors": "^1.0.0",
+    "react-transform-hmr": "^1.0.1",
+    "redbox-react": "^1.1.1",
+    "sass-loader": "^3.1.1",
     "script-loader": "^0.6.1",
+    "style-loader": "^0.12.4",
+    "svg-inline-loader": "^0.3.0",
+    "svgo-loader": "^1.1.0",
+    "webpack": "^1.12.3",
+    "webpack-dev-server": "^1.12.1",
     "worker-loader": "^0.6.0"
   },
   "jest": {
