@@ -4,6 +4,7 @@ import Immutable from "immutable";
 import AccountCard from "./AccountCard";
 import RecentTransactions from "../Account/RecentTransactions";
 import Translate from "react-translate-component";
+import Proposals from "components/Account/Proposals";
 
 class Dashboard extends Component {
 
@@ -17,7 +18,10 @@ class Dashboard extends Component {
         for(let a of names)
             itemRows.push(<AccountCard key={a} account={a}/>);
 
-        return (
+        return (<div>
+            {/* P R O P O S A L S 
+            <Proposals/>
+            <br/>*/}
             <div className="grid-block page-layout vertical medium-horizontal">
                 <div className="grid-block medium-8 flex-start" style={{overflowY: "auto", zIndex: 1}}>
                     <div className="grid-block regular-padding small-up-1 medium-up-2 large-up-3">
@@ -31,7 +35,7 @@ class Dashboard extends Component {
                     </div>
                 </div>
             </div>
-        );
+        </div>);
     }
 }
 
